Removed the legacy predefined shared variables: "html_escape", "normalize_newlines", "xml_escape", "capture_output", "compress". It had to be done now as TemplateTransformModel will be removed as part of FREEMARKER-63.
