+++ freemarker-core-test/src/test/resources/org/apache/freemarker/core/templatesuite/expected/compress.txt
@@ -28,10 +28,6 @@
 
 <p>This is the same message, using the &quot;compress&quot; tag:</p>
 <p>Hello, world!</p>
-<p>This is the same message, using the &quot;StandardCompress&quot; transform model:</p>
-<p>Hello, world!</p>
-<p>This multi-line message should be compressed to a single line.</p>
-
 <p>An example where the first character is not whitespace but the second character is:</p>
 <p>x y</p>
