+++ freemarker-core-test/src/test/resources/org/apache/freemarker/core/templatesuite/templates/compress.ftl
@@ -22,7 +22,6 @@
 </head>
 <body>
 
-<#assign utility={'standardCompress': "org.apache.freemarker.core.util.StandardCompress"?new()}>
 <p>A simple test follows:</p>
 
 <p>${message}</p>
@@ -34,22 +33,6 @@
 
 <p>${message}</p>
 </#compress>
-
-<@utility.standardCompress buffer_size=8>
-
-  <p>This is the same message,  using the &quot;StandardCompress&quot; transform model:</p>
-
-
-<p>${message}</p>
-</@>
-
-<@utility.standardCompress single_line=true>
-
-<p>This 
-   multi-line message 
-     should 
-be compressed 
-   to a single line.</p></@>
 
 <p>An example where the first character is not whitespace but the second character is:</p>
 <p><#compress>x y</#compress></p>
